# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Accidents",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://localhost:8000",
            "description": "accident de 2021 en france"
        }
    ],
    "paths": {
        "/api/caracteristiques/new": {
            "post": {
                "tags": [
                    "Media"
                ],
                "summary": "Upload document",
                "description": "",
                "operationId": "6087348bcb41022552c65d6072be8d1d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/octet-stream": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "description": "Binary content of file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/caracteristiques": {
            "get": {
                "operationId": "5728a9cf244bcce91bd269fcba11245d",
                "responses": {
                    "200": {
                        "description": "Liste des caracteristiques",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Liste des caracteristiques",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CaracteristiqueDisplayOnMap"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/caracteristiques/num_acc/{num_acc}": {
            "get": {
                "tags": [
                    "num_acc"
                ],
                "description": "caracteristiques par numéro d'accident",
                "operationId": "find carac by num_acc",
                "parameters": [
                    {
                        "name": "num_acc",
                        "in": "path",
                        "description": "num_acc of this item",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid value"
                    }
                },
                "deprecated": false
            }
        },
        "/api/caracteristiques/paginated/{row_index}": {
            "get": {
                "tags": [
                    "row_index"
                ],
                "description": "50 caracteristiques par numéro row_index",
                "operationId": "find carac by row_index",
                "parameters": [
                    {
                        "name": "row_index",
                        "in": "path",
                        "description": "row_index of this item",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid value"
                    }
                },
                "deprecated": false
            }
        },
        "/api/caracteristiques/paginated_csv/{row_index}": {
            "get": {
                "tags": [
                    "row_index"
                ],
                "description": "50 caracteristiques en csv par numéro row_index",
                "operationId": "find carac csv by row_index",
                "parameters": [
                    {
                        "name": "row_index",
                        "in": "path",
                        "description": "row_index of this item",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid value"
                    }
                },
                "deprecated": false
            }
        },
        "/api/caracteristiques/paginated_json/{row_index}": {
            "get": {
                "tags": [
                    "row_index"
                ],
                "description": "50 caracteristiques json par numéro row_index",
                "operationId": "find carac json by row_index",
                "parameters": [
                    {
                        "name": "row_index",
                        "in": "path",
                        "description": "row_index of this item",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid value"
                    }
                },
                "deprecated": false
            }
        },
        "/api/caracteristiques/add_c": {
            "post": {
                "tags": [
                    "caracteristique"
                ],
                "description": "Add a new caracteristique",
                "operationId": "addCaracteristique",
                "requestBody": {
                    "description": "Caracteristique object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/CaracteristiqueSingle"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CaracteristiqueSingle"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "deprecated": false
            }
        }
    },
    "components": {
        "schemas": {
            "CaracteristiqueSingle": {
                "title": "Caracteristique all fields",
                "description": "Caracteristique des accidents",
                "properties": {
                    "id": {
                        "description": "Identifiant de la caracteristique",
                        "type": "integer"
                    },
                    "num_acc": {
                        "description": "Numéro d'accident",
                        "type": "integer"
                    },
                    "jour": {
                        "description": "jour de l'accident",
                        "type": "integer"
                    },
                    "mois": {
                        "description": "mois de l'accident",
                        "type": "integer"
                    },
                    "an": {
                        "description": "année de l'accident",
                        "type": "integer"
                    },
                    "hrmn": {
                        "description": "heure de l'accident",
                        "type": "string"
                    },
                    "lum": {
                        "description": "lumière au moment de l'accident",
                        "type": "integer"
                    },
                    "agg": {
                        "description": "dans agglomeration/hors agglomeration",
                        "type": "integer"
                    },
                    "int": {
                        "description": "intersection de l'accident",
                        "type": "integer"
                    },
                    "atm": {
                        "description": "condition metéos de l'accident",
                        "type": "integer"
                    },
                    "col": {
                        "description": "type de collision de l'accident",
                        "type": "integer"
                    },
                    "com": {
                        "description": "code commune de l'accident",
                        "type": "integer"
                    },
                    "dep": {
                        "description": "code département de l'accident",
                        "type": "integer"
                    },
                    "adr": {
                        "description": "adresse postale",
                        "type": "string"
                    },
                    "lat": {
                        "description": "latitude",
                        "type": "integer"
                    },
                    "longi": {
                        "description": "longitude",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "CaracteristiqueDisplayOnMap": {
                "title": "Caracteristique for display on a map",
                "description": "Caracteristique des accidents",
                "properties": {
                    "id": {
                        "description": "Identifiant de la caracteristique",
                        "type": "integer"
                    },
                    "num_acc": {
                        "description": "Numéro d'accident",
                        "type": "integer"
                    },
                    "lum": {
                        "description": "lumière au moment de l'accident",
                        "type": "integer"
                    },
                    "lat": {
                        "description": "latitude",
                        "type": "integer"
                    },
                    "longi": {
                        "description": "longitude",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    }
}